requirements.txt
discord
mysql.connector
asyncio
datetime
schedule
